pyproject.toml
[project]
name = "templatrix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "SaiDhinakar", email = "saidhinakars899@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[all,standard]>=0.115.13",
    "flask>=3.1.1",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "setuptools>=80.9.0",
    "uvicorn>=0.34.3",
]

[project.scripts]
templatrix = "templatrix.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
